client: Move opts to separate files
